Move flight commands to their own package
